slider status

+++ app/Http/Controllers/Admin/UserController.php
@@ -8,7 +8,9 @@ use App\Model\User;
 use App\Model\Role;
 use App\Model\FfwcStation;
 use App\Model\UserStation;
+use App\Model\SlideDetail;
 use App\Model\Location;
+use App\Model\SlideStatus;
 use Hash;
 
 
@@ -35,7 +37,8 @@ class UserController extends Controller
         $roles = Role::all();
         $locations = Location::all();
         $ffwcStations = FfwcStation::all();
-        return view('admin/user/create',compact('roles','locations','ffwcStations'));
+        $slideDetails = SlideDetail::all();
+        return view('admin/user/create',compact('roles','locations','ffwcStations','slideDetails'));
     }
 
     /**
@@ -49,10 +52,10 @@ class UserController extends Controller
         // dd($request);
         $validateData = $request->validate([
             'role_id' => 'required',
-            'location_id' => 'required',
+            // 'location_id' => 'required',
             'email' => 'required',
-            'ffwc_sations' => 'required',
-            'zoom_level' => 'required',
+            // 'ffwc_sations' => 'required',
+            // 'zoom_level' => 'required',
             'password' => 'required',
 
         ]);
@@ -64,17 +67,67 @@ class UserController extends Controller
         $user->role_id = $request->input('role_id');
         $user->location_id = $request->input('location_id');
         $user->zoom_level = $request->input('zoom_level');
-        $user->user_slide_description = $request->input('user_slide_description');
         $user->save();
         $userId = $user->id;
-        $ffwc_sations = $request->input('ffwc_sations');
-        foreach($ffwc_sations as $ffwc_sation){
-            $userStation = new UserStation();
-            $userStation->user_id = $userId;
-            $userStation->ffwc_stations_id = $ffwc_sation;
-            $userStation->save();
+        if($request->input('ffwc_sations')){
+            $ffwc_sations = $request->input('ffwc_sations');
+            foreach($ffwc_sations as $ffwc_sation){
+                $userStation = new UserStation();
+                $userStation->user_id = $userId;
+                $userStation->ffwc_stations_id = $ffwc_sation;
+                $userStation->save();
+            }
         }
+        // if($request->input('$slide_1')) {
+        //     $slide_1 = $request->input('$slide_1');
+        // }
+        // else{
+        //     $slide_1 = '';
+        // }
+
+
+        // if($request->input('$slide_2')) {
+        //     $slide_2 = $request->input('$slide_2');
+        // }
+        // else{
+        //     $slide_2 = '';
+        // }
+        // if($request->input('$slide_3')) {
+        //     $slide_3 = $request->input('$slide_3');
+        // }
+        // else{
+        //     $slide_3 = '';
+        // }
+
+
+        // if($request->input('$slide_4')) {
+        //     $slide_4 = $request->input('$slide_4');
+        // }
+        // else{
+        //     $slide_4 = '';
+        // }
+        // if($request->input('$slide_5')) {
+        //     $slide_5 = $request->input('$slide_5');
+        // }
+        // else{
+        //     $slide_5 = '';
+        // }
+        // if($request->input('$slide_6')) {
+        //     $slide_6 = $request->input('$slide_6');
+        // }
+        // else{
+        //     $slide_6 = '';
+        // }
+        // if($request->input('$slide_7')) {
+        //     $slide_7 = $request->input('$slide_7');
+        // }
+        // else{
+        //     $slide_7= '';
+        // }
+
+        // $sort = 
         
+       
          return redirect()->route('userlist')->with('message','User Created Successfully!'); 
     }
 
@@ -100,10 +153,11 @@ class UserController extends Controller
         $roles = Role::all();
         $locations = Location::all();
         $ffwcStations = FfwcStation::all();
+        $slideDetails = SlideDetail::all();
         $userStations = UserStation::where('user_id',$id)->pluck('ffwc_stations_id')->toArray();
         // dd((array)$userStations);
         $user = User::Where('id',$id)->first();
-        return view('admin/user/edit', compact('user','roles','locations','ffwcStations','userStations'));
+        return view('admin/user/edit', compact('user','roles','locations','ffwcStations','userStations','slideDetails'));
     }
 
     /**
@@ -118,10 +172,10 @@ class UserController extends Controller
         $validateData = $request->validate([
             
             'role_id' => 'required',
-            'location_id' => 'required',
+            // 'location_id' => 'required',
             'email' => 'required',
-            'ffwc_sations' => 'required',
-            'zoom_level' => 'required',
+            // 'ffwc_sations' => 'required',
+            // 'zoom_level' => 'required',
         ]);
         $id = $request->input('id');
         $user =  User::where('id',$id)->first();
@@ -132,18 +186,18 @@ class UserController extends Controller
         $user->role_id = $request->input('role_id');
         $user->location_id = $request->input('location_id');
         $user->zoom_level = $request->input('zoom_level');
-        $user->user_slide_description = $request->input('user_slide_description');
         $user->save();
-
-        $ffwc_sations = $request->input('ffwc_sations');
-        $userStation = UserStation::where('user_id',$id)->delete();
-        foreach($ffwc_sations as $ffwc_sation){
-            $userStation = new UserStation();
-            $userStation->user_id = $id;
-            $userStation->ffwc_stations_id = $ffwc_sation;
-            $userStation->save();
+        if($request->input('ffwc_sations')){
+            $ffwc_sations = $request->input('ffwc_sations');
+            $userStation = UserStation::where('user_id',$id)->delete();
+            foreach($ffwc_sations as $ffwc_sation){
+                $userStation = new UserStation();
+                $userStation->user_id = $id;
+                $userStation->ffwc_stations_id = $ffwc_sation;
+                $userStation->save();
+            }
         }
-
+    
          return redirect()->route('userlist')->with('message','User Updated Successfully!'); 
 
     }
